+++ chapter_001/src/test/java/ru/job4j/calculator/ConverterTest.java
@@ -8,28 +8,28 @@ public class ConverterTest {
     @Test
     public void when60RubleToDollarThen1() {
         Converter converter = new Converter();
-        int result = converter.rubleToDollar(60);
+        double result = converter.rubleToDollar(60);
         assertThat(result, is(1D));
     }
 
     @Test
     public void when70RubleToEuroThen1() {
         Converter converter = new Converter();
-        int result = converter.rubleToEuro(70);
+        double result = converter.rubleToEuro(70);
         assertThat(result, is(1D));
     }
 
     @Test
     public void when70EuroToRuble() {
         Converter converter = new Converter();
-        int result = converter.euroToRuble(70);
+        double result = converter.euroToRuble(70);
         assertThat(result, is(4900D));
     }
 
     @Test
     public void when60DollarsToRuble() {
         Converter converter = new Converter();
-        int result = converter.dollarToRuble(60);
+        double result = converter.dollarToRuble(60);
         assertThat(result, is(3600D));
     }
 }
